refactor: change the type of timestampMillisec to long

+++ Assets/MediaPipeUnity/Samples/Scenes/Tasks/Face Detection/FaceDetectorRunner.cs
@@ -124,7 +124,7 @@ namespace Mediapipe.Unity.Sample.FaceDetection
       }
     }
 
-    private void OnFaceDetectionsOutput(FaceDetectionResult result, Image image, int timestamp)
+    private void OnFaceDetectionsOutput(FaceDetectionResult result, Image image, long timestamp)
     {
       _detectionResultAnnotationController.DrawLater(result);
     }
